skills/video-pipeline/requirements.txt
# Video Production Pipeline Dependencies

# AI & LLM
anthropic>=0.39.0

# Database
pyairtable>=2.3.0

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Slack
slack-sdk>=3.27.0

# Voice Synthesis
elevenlabs>=1.0.0

# Image Generation (uncomment what you use)
openai>=1.12.0  # For DALL-E
# fal-client>=0.4.0  # For Fal.ai

# YouTube
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.2

# Web Scraping
apify-client>=1.6.0  # YouTube trending scraper

# Research Agent - Web Search
tavily-python>=0.3.0  # Tavily search API for research scanning

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Audio Processing
mutagen>=1.47.0  # Audio duration detection

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.0  # Retry logic
rich>=13.7.0  # Pretty console output
